Refactor modal iconColor to type
