+++ frontend/src/citizen-frontend/income-statements/IncomeStatements.tsx
@@ -181,7 +181,7 @@ export default React.memo(function IncomeStatements() {
 
           {deletionState.status !== 'row-not-selected' && (
             <InfoModal
-              iconColor="orange"
+              type="warning"
               title={t.income.table.deleteConfirm}
               text={t.income.table.deleteDescription}
               icon={faQuestion}
