+++ frontend/src/employee-frontend/components/MobilePairingModal.tsx
@@ -1,4 +1,4 @@
-// SPDX-FileCopyrightText: 2017-2021 City of Espoo
+// SPDX-FileCopyrightText: 2017-2022 City of Espoo
 //
 // SPDX-License-Identifier: LGPL-2.1-or-later
 
@@ -127,7 +127,7 @@ export default React.memo(function MobilePairingModal({
       }
       icon={faPlus}
       data-qa={`mobile-pairing-modal-phase-${phase}`}
-      iconColor={phase === 3 ? 'green' : 'blue'}
+      type={phase === 3 ? 'success' : 'info'}
       {...actions}
     >
       {phase === 1 && (
