+++ frontend/src/employee-frontend/components/PersonalMobileDevicesPage.tsx
@@ -138,7 +138,7 @@ const EditNameModal = React.memo(function EditNameModal({
     <InfoModal
       title={i18n.personalMobileDevices.editName}
       icon={faQuestion}
-      iconColor="blue"
+      type="info"
       reject={{ action: close, label: i18n.common.cancel }}
       resolve={{
         action: () => putMobileDeviceName(id, newName).then(close),
@@ -170,7 +170,7 @@ const DeleteModal = React.memo(function DeleteModal({
     <InfoModal
       title={i18n.personalMobileDevices.deleteDevice}
       icon={faQuestion}
-      iconColor="orange"
+      type="warning"
       reject={{ action: close, label: i18n.common.cancel }}
       resolve={{
         action: () => deleteMobileDevice(id).then(close),
