+++ frontend/src/employee-frontend/components/child-information/FeeAlteration.tsx
@@ -104,7 +104,7 @@ export default React.memo(function FeeAlteration({ id, startOpen }: Props) {
       </CollapsibleContentArea>
       {deleted ? (
         <InfoModal
-          iconColor="orange"
+          type="warning"
           title={i18n.childInformation.feeAlteration.confirmDelete}
           text={`${
             i18n.childInformation.feeAlteration.types[deleted.type]
