+++ frontend/src/employee-frontend/components/child-information/backup-care/BackupCareRow.tsx
@@ -57,7 +57,7 @@ export default function BackupCareRow({ childId, backupCare }: Props) {
     <div>
       {uiMode === `remove-backup-care-${backupCare.id}` && (
         <InfoModal
-          iconColor="orange"
+          type="warning"
           title={i18n.childInformation.backupCares.remove}
           text={`${
             i18n.common.period
