+++ frontend/src/employee-frontend/components/fee-decision-details/FeeDecisionDetailsPage.tsx
@@ -115,7 +115,7 @@ export default React.memo(function FeeDecisionDetailsPage() {
       {confirmingBack && (
         <InfoModal
           title={i18n.feeDecision.modal.title}
-          iconColor="orange"
+          type="warning"
           icon={faQuestion}
           reject={{ action: goBack, label: i18n.feeDecision.modal.cancel }}
           resolve={{
