+++ frontend/src/employee-frontend/components/invoices/InvoicesPage.tsx
@@ -1,4 +1,4 @@
-// SPDX-FileCopyrightText: 2017-2020 City of Espoo
+// SPDX-FileCopyrightText: 2017-2022 City of Espoo
 //
 // SPDX-License-Identifier: LGPL-2.1-or-later
 
@@ -104,7 +104,7 @@ const Modal = React.memo(function Modal({
 
   return (
     <AsyncFormModal
-      iconColor="blue"
+      type="info"
       title={i18n.invoices.sendModal.title}
       icon={faEnvelope}
       resolveAction={() => sendInvoices({ invoiceDate, dueDate })}
