+++ frontend/src/employee-frontend/components/person-profile/PersonFeeDecisions.tsx
@@ -1,4 +1,4 @@
-// SPDX-FileCopyrightText: 2017-2021 City of Espoo
+// SPDX-FileCopyrightText: 2017-2022 City of Espoo
 //
 // SPDX-License-Identifier: LGPL-2.1-or-later
 
@@ -143,7 +143,7 @@ const Modal = React.memo(function Modal({
   return (
     <AsyncFormModal
       icon={faPlus}
-      iconColor="blue"
+      type="info"
       title={i18n.personProfile.feeDecisions.createRetroactive}
       resolveAction={resolve}
       resolveLabel={i18n.common.create}
