+++ frontend/src/employee-frontend/components/person-profile/PersonFridgePartner.tsx
@@ -76,7 +76,7 @@ const PersonFridgePartner = React.memo(function PersonFridgePartner({
         />
       ) : uiMode === `remove-fridge-partner-${selectedPartnershipId}` ? (
         <InfoModal
-          iconColor="orange"
+          type="warning"
           title={i18n.personProfile.fridgePartner.removePartner}
           text={i18n.personProfile.fridgePartner.confirmText}
           icon={faQuestion}
