+++ frontend/src/employee-frontend/components/person-profile/person-fridge-partner/FridgePartnerModal.tsx
@@ -1,4 +1,4 @@
-// SPDX-FileCopyrightText: 2017-2020 City of Espoo
+// SPDX-FileCopyrightText: 2017-2022 City of Espoo
 //
 // SPDX-License-Identifier: LGPL-2.1-or-later
 
@@ -139,7 +139,7 @@ function FridgePartnerModal({ partnership, onSuccess, headPersonId }: Props) {
           : i18n.personProfile.fridgePartner.newPartner
       }
       icon={partnership ? faPen : faUser}
-      iconColor="blue"
+      type="info"
       resolveAction={onSubmit}
       resolveLabel={i18n.common.confirm}
       resolveDisabled={!form.partner || validationErrors.length > 0}
