+++ frontend/src/employee-frontend/components/unit/tab-groups/groups/group/GroupTransferModal.tsx
@@ -1,4 +1,4 @@
-// SPDX-FileCopyrightText: 2017-2020 City of Espoo
+// SPDX-FileCopyrightText: 2017-2022 City of Espoo
 //
 // SPDX-License-Identifier: LGPL-2.1-or-later
 
@@ -114,7 +114,7 @@ export default React.memo(function GroupTransferModal({
       data-qa="group-placement-modal"
       title={i18n.unit.placements.modal.transferTitle}
       icon={faExchange}
-      iconColor="blue"
+      type="info"
       resolveAction={submitForm}
       resolveLabel={i18n.common.confirm}
       resolveDisabled={form.errors.length > 0}
