+++ frontend/src/employee-frontend/components/unit/tab-groups/missing-group-placements/BackupCareGroupModal.tsx
@@ -1,4 +1,4 @@
-// SPDX-FileCopyrightText: 2017-2020 City of Espoo
+// SPDX-FileCopyrightText: 2017-2022 City of Espoo
 //
 // SPDX-License-Identifier: LGPL-2.1-or-later
 
@@ -71,7 +71,7 @@ export default React.memo(function BackupCareGroupModal({
           : i18n.unit.placements.modal.createTitle
       }
       icon={isTransfer ? faExchange : faChild}
-      iconColor="blue"
+      type="info"
       resolveAction={submitForm}
       resolveLabel={i18n.common.confirm}
       resolveDisabled={!group}
